+++ src/Libraries/SmartStore.Core/Email/IEmailSender.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace SmartStore.Core.Email
+{    
+    /// <summary>
+    /// contract for email sender 
+    /// </summary>
+    public interface IEmailSender
+    {
+        void SendEmail(SmtpContext smtpContext, EmailMessage message);
+		Task SendEmailAsync(SmtpContext smtpContext, EmailMessage message); 
+    }
+}