+++ src/Libraries/SmartStore.Services/Messages/QueuedMessagesSendTask.cs
@@ -1,4 +1,8 @@
 ﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using SmartStore.Core.Email;
 using SmartStore.Services.Logging;
 using SmartStore.Services.Tasks;
 
@@ -11,15 +15,19 @@ namespace SmartStore.Services.Messages
     {
         private readonly IQueuedEmailService _queuedEmailService;
         private readonly IEmailSender _emailSender;
-        private readonly ILogger _logger;
 
-        public QueuedMessagesSendTask(IQueuedEmailService queuedEmailService,
-            IEmailSender emailSender, ILogger logger)
+        public QueuedMessagesSendTask(IQueuedEmailService queuedEmailService, IEmailSender emailSender)
         {
             this._queuedEmailService = queuedEmailService;
             this._emailSender = emailSender;
-            this._logger = logger;
+			Logger = NullLogger.Instance;
         }
+
+		public ILogger Logger
+		{
+			get;
+			set;
+		}
 
         /// <summary>
         /// Executes a task
@@ -27,32 +35,50 @@ namespace SmartStore.Services.Messages
         public void Execute()
         {
             var maxTries = 3;
-            var queuedEmails = _queuedEmailService.SearchEmails(null, null, null, null,
-                true, maxTries, false, 0, 10000);
-            foreach (var queuedEmail in queuedEmails)
+            var queuedEmails = _queuedEmailService.SearchEmails(null, null, null, null, true, maxTries, false, 0, 10000);
+
+            foreach (var qe in queuedEmails)
             {
-                var bcc = String.IsNullOrWhiteSpace(queuedEmail.Bcc) 
+                var bcc = String.IsNullOrWhiteSpace(qe.Bcc) 
                             ? null 
-                            : queuedEmail.Bcc.Split(new char[] { ';' }, StringSplitOptions.RemoveEmptyEntries);
-                var cc = String.IsNullOrWhiteSpace(queuedEmail.CC) 
+                            : qe.Bcc.Split(new char[] { ';' }, StringSplitOptions.RemoveEmptyEntries);
+                var cc = String.IsNullOrWhiteSpace(qe.CC) 
                             ? null 
-                            : queuedEmail.CC.Split(new char[] { ';' }, StringSplitOptions.RemoveEmptyEntries);
+                            : qe.CC.Split(new char[] { ';' }, StringSplitOptions.RemoveEmptyEntries);
 
                 try
                 {
-                    _emailSender.SendEmail(queuedEmail.EmailAccount, queuedEmail.Subject, queuedEmail.Body,
-                       queuedEmail.From, queuedEmail.FromName, queuedEmail.To, queuedEmail.ToName, bcc, cc);
+					var smtpContext = new SmtpContext(qe.EmailAccount);
 
-                    queuedEmail.SentOnUtc = DateTime.UtcNow;
+					var msg = new EmailMessage(
+						new EmailAddress(qe.To, qe.ToName),
+						qe.Subject,
+						qe.Body,
+						new EmailAddress(qe.From, qe.FromName));
+
+					if (qe.ReplyTo.HasValue()) 
+					{
+						msg.ReplyTo.Add(new EmailAddress(qe.ReplyTo, qe.ReplyToName));
+					}
+
+					if (cc != null)
+						msg.Cc.AddRange(cc.Where(x => x.HasValue()).Select(x => new EmailAddress(x)));
+
+					if (bcc != null)
+						msg.Bcc.AddRange(bcc.Where(x => x.HasValue()).Select(x => new EmailAddress(x)));
+
+					_emailSender.SendEmail(smtpContext, msg);
+
+                    qe.SentOnUtc = DateTime.UtcNow;
                 }
                 catch (Exception exc)
                 {
-                    _logger.Error(string.Format("Error sending e-mail. {0}", exc.Message), exc);
+					Logger.Error(string.Format("Error sending e-mail: {0}", exc.Message), exc);
                 }
                 finally
                 {
-                    queuedEmail.SentTries = queuedEmail.SentTries + 1;
-                    _queuedEmailService.UpdateQueuedEmail(queuedEmail);
+                    qe.SentTries = qe.SentTries + 1;
+                    _queuedEmailService.UpdateQueuedEmail(qe);
                 }
             }
         }